habitv6

+++ src/main/java/spring_group1/com/services/serviceImpl/HabitServiceImpl.java
@@ -2,6 +2,8 @@ package spring_group1.com.services.serviceImpl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import spring_group1.com.exception.DuplicateName;
+import spring_group1.com.exception.NotFoundExceptionHandler;
 import spring_group1.com.model.Habit;
 import spring_group1.com.repository.HabitRepository;
 import spring_group1.com.request.HabitRequest;
@@ -17,26 +19,59 @@ public class HabitServiceImpl implements HabitService {
 
     @Override
     public List<Habit> getAllHabit(Integer page, Integer size) {
+        if(page <= 0 ){
+            throw new NotFoundExceptionHandler("Page cannot negative or zero");
+        }
+        if(size <= 0) {
+            throw new NotFoundExceptionHandler("Size cannot negative or zero");
+        }
         return habitRepository.findAllHabit(page, size);
     }
 
     @Override
     public Habit getHabitById(Integer habitId) {
-        return habitRepository.findHabitById(habitId);
+        if (habitRepository.findHabitById(habitId) == null) {
+            throw new NotFoundExceptionHandler("Not found Habit ID "+ habitId);
+        }else {
+            return habitRepository.findHabitById(habitId);
+        }
     }
 
     @Override
     public Habit createhabit(HabitRequest habitRequest) {
+        List<Habit> habits = habitRepository.getAllHabit();
+
+        for(Habit hab : habits){
+            if(habitRequest.getTitle() != null){
+                throw new DuplicateName("This habit already has!");
+            }
+        }
+
         return habitRepository.createNewHabit(habitRequest);
     }
 
     @Override
     public Habit updateHabit(Integer habitId, HabitRequest habitRequest) {
+        if (habitRepository.findHabitById(habitId) == null) {
+            throw new NotFoundExceptionHandler("Not found Habit ID "+ habitId);
+        }
+        List<Habit> habits = habitRepository.getAllHabit();
+
+        for(Habit hab : habits){
+            if(habitRequest.getTitle() != null){
+                throw new DuplicateName("This habit already has!");
+            }
+        }
         return habitRepository.updateNewHabit(habitId, habitRequest);
     }
 
     @Override
     public boolean deleteHabit(Integer habitId) {
+
+        if (habitRepository.findHabitById(habitId) == null) {
+            throw new NotFoundExceptionHandler("Not found Habit ID "+ habitId);
+        }
+
         int rowEffect = habitRepository.deleteHabitById(habitId);
         return rowEffect > 0;
     }
